feat: refactor autotasks, add better logs
