+++ autotasks/aave-harvester/src/index.ts
@@ -1,4 +1,4 @@
 import { makeHandler } from "../../handler";
-import { aaveHarvester } from "./aave-harvester";
+import { aaveHarvester } from "../aave-harvester";
 
 export const handler = makeHandler(aaveHarvester);
